\name{runApp.xmsPANDA}
\alias{runApp.xmsPANDA}
\title{runApp.xmsPANDA}
\description{
Funtion to run the full xmsPANDA Shiny App
}
\usage{
runApp.xmsPANDA()
}
\details{
This function runs the Shiny App which includes capabilities for replicate summarization,
data preprocessing, feature selection, PCA, HCA, classification accuracy evaluation,
differential network analysis, module preservation analysis,
global clustering analysis, functional class scoring, comparing normalization methods,
and metabolite quantification analysis.
}
\value{}
\author{
Karan Uppal (kuppal3gt@gmail.com)
}
