\name{runApp.xmsPANDALite}
\alias{runApp.xmsPANDALite}
\title{runApp.xmsPANDALite}
\description{
Funtion to run the Shiny App for the Lite version of xmsPANDA
}
\usage{
runApp.xmsPANDALite()
}
\details{
This function runs the Lite Shiny App which includes capabilities for 
data preprocessing, feature selection, network analysis, PCA, HCA,
and classification accuracy evaluation.
}
\value{}
\author{
Karan Uppal (kuppal3gt@gmail.com)
}
