\name{xmsPANDA-package}
\alias{xmsPANDA-package}
\alias{xmsPANDA}
\docType{package}
\title{xmsPANDA}
\description{R pacakge for biomarker discovery, supervised and unsupervised learning, and network analysis.}
\details{
\tabular{ll}{
Package: \tab xmsPANDA\cr
Type: \tab Package\cr
Version: \tab 1.1.75\cr
Date: \tab 2020-12-11\cr
License: \tab gpl2.0\cr
LazyLoad: \tab yes\cr
}
}
\author{
Karan Uppal
Maintainer: <kuppal2@emory.edu>
}
\keyword{ xmsPANDA }
